// src/shared/interacts.enum.ts
export enum InteractType {
  LOVE = 'LOVE',
  DISLOVE = 'DISLOVE',
  // SAVE = 'SAVE',
  // VIEW = 'VIEW',
  // SHARE = 'SHARE',
  COMMENT = 'COMMENT',
  REPLYCOMMENT = 'REPLYCOMMENT',

  // HIDE = 'HIDE',
  // PIN = 'PIN',
  // UNPIN = 'UNPIN',
  // TURNON = 'TURNON',
  // TURNOFF = 'TURNOFF',

  // HIDEUSER = 'HIDEUSER',
}

export enum EntityInteractType {
  POST = 'Post',
  COMMENT = 'Comment',
  USER = 'User',
  CONVERSATION = 'Conversation',
}